requirements.txt
asgiref==3.2.10
buttercms-python==1.2
certifi==2020.6.20
chardet==3.0.4
Django==3.0.9
django-active-link==0.1.6
idna==2.10
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
